+++ ciron/crypto.h
@@ -17,7 +17,7 @@ extern "C" {
  *
  * The result will not be \0 terminated.
  */
-CironError CIRONAPI ciron_generate_salt(CironContext context, int nbytes,
+CironError CIRONAPI ciron_generate_salt(CironContext context, size_t nbytes,
 		unsigned char *buf);
 
 
@@ -30,7 +30,7 @@ CironError CIRONAPI ciron_generate_salt(CironContext context, int nbytes,
  *
  * The result will not be \0 terminated.
  */
-CironError CIRONAPI ciron_generate_iv(CironContext context, int nbytes,
+CironError CIRONAPI ciron_generate_iv(CironContext context, size_t nbytes,
 		unsigned char *buf);
 
 
@@ -50,9 +50,9 @@ CironError CIRONAPI ciron_generate_iv(CironContext context, int nbytes,
  * The result will not be \0 terminated.
  */
 CironError CIRONAPI ciron_generate_key(CironContext context,
-		const unsigned char* password, int password_len,
-		const unsigned char *salt, int salt_len, CironAlgorithm algorithm,
-		int iterations, unsigned char *buf);
+		const unsigned char* password, size_t password_len,
+		const unsigned char *salt, size_t salt_len, CironAlgorithm algorithm,
+		unsigned int iterations, unsigned char *buf);
 
 /** Encrypt the provided data using the specified algorithm.
  *
@@ -81,7 +81,7 @@ CironError CIRONAPI ciron_generate_key(CironContext context,
  */
 CironError CIRONAPI ciron_encrypt(CironContext context, CironAlgorithm algorithm,
 		const unsigned char *key, const unsigned char *iv,
-		const unsigned char *data, int data_len, unsigned char *buf, int *sizep);
+		const unsigned char *data, size_t data_len, unsigned char *buf, size_t *sizep);
 
 /** Decrypt the provided data using the specified algorithm.
  *
@@ -110,7 +110,7 @@ CironError CIRONAPI ciron_encrypt(CironContext context, CironAlgorithm algorithm
  */
 CironError CIRONAPI ciron_decrypt(CironContext context, CironAlgorithm algorithm,
 		const unsigned char *key, const unsigned char *iv,
-		const unsigned char *data, int data_len, unsigned char *buf, int *sizep);
+		const unsigned char *data, size_t data_len, unsigned char *buf, size_t *sizep);
 
 
 /** Calculates an HMAC from the provided data using password, salt,
@@ -132,10 +132,10 @@ CironError CIRONAPI ciron_decrypt(CironContext context, CironAlgorithm algorithm
  *
  */
 CironError CIRONAPI ciron_hmac(CironContext context, CironAlgorithm algorithm,
-		const unsigned char *password, int password_len,
-		const unsigned char *salt_bytes, int salt_len, int iterations,
-		const unsigned char *data, int data_len, unsigned char *result,
-		int *result_len);
+		const unsigned char *password, size_t password_len,
+		const unsigned char *salt_bytes, size_t salt_len, unsigned int iterations,
+		const unsigned char *data, size_t data_len, unsigned char *result,
+		size_t *result_len);
 
 #ifdef __cplusplus
 } // extern "C"
